requirements.txt
tensorflow==2.15.0
opencv-python
numpy
matplotlib
seaborn
scikit-learn
Pillow
streamlit
